feat: move Cli & Argument to Confirma.Console namespace
